requirements.txt
langchain>=0.3.25
langchain-community>=0.3.24
langchain-core>=0.3.59
langgraph>=0.2.45
langchain-groq>=0.3.2          
langchain-openai>=0.3.16       
chromadb>=0.5.23
sentence-transformers>=3.3.1
openai>=1.82.0                
groq>=0.11.0
python-dotenv>=1.0.1
pydantic>=2.10.6
